zmiana zakresu zmiennych

+++ src/main/java/core/basesyntax/Circle.java
@@ -1,18 +1,18 @@
 package core.basesyntax;
 
 public class Circle extends Figure {
-    private int radius = 0;
+    private double radius = 0;
     private String name = "circle";
     private String color = "white";
     private double area = 0; 
 
-    public Circle(int radius, String color) {
+    public Circle(double radius, String color) {
         this.radius = radius;
         this.color = color;
         this.area = this.getArea(radius);
     }
     
-    public double getArea(int radius) {
+    public double getArea(double radius) {
         double area = Math.PI * radius * radius;
         return area;
     }
