+++ src/main/java/core/basesyntax/IsoscelesTrapezoid.java
@@ -1,14 +1,14 @@
 package core.basesyntax;
 
 public class IsoscelesTrapezoid extends Figure {
-    private int side1;
-    private int side2;
-    private int sideh;
+    private double side1;
+    private double side2;
+    private double sideh;
     private String name = "isosceles trapezoid";
     private String color = "";
     private double area = 0;
 
-    public IsoscelesTrapezoid(int side1, int side2, int sideh, String color) {
+    public IsoscelesTrapezoid(double side1, double side2, double sideh, String color) {
         this.side1 = side1;
         this.side2 = side2;
         this.sideh = sideh;
@@ -16,7 +16,7 @@ public class IsoscelesTrapezoid extends Figure {
         this.area = this.getArea(side1, side2, sideh);
     }
 
-    public double getArea(int side1, int side2, int sideh) {
+    public double getArea(double side1, double side2, double sideh) {
         return (side1 + side2) * sideh / 2;
     }
