+++ src/main/java/core/basesyntax/RightTriangle.java
@@ -1,20 +1,20 @@
 package core.basesyntax;
 
 public class RightTriangle extends Figure {
-    private int firstLeg = 0;
-    private int secondLeg;
+    private double firstLeg = 0;
+    private double secondLeg;
     private String name = "right triangle";
     private String color = "";
     private double area = 0;
 
-    public RightTriangle(int firstLeg, int secondLeg, String color) {
+    public RightTriangle(double firstLeg, double secondLeg, String color) {
         this.firstLeg = firstLeg;
         this.secondLeg = secondLeg;
         this.color = color;
         this.area = this.getArea(firstLeg, secondLeg);
     }
 
-    public double getArea(int firstLeg, int secondLeg) {
+    public double getArea(double firstLeg, double secondLeg) {
         double area = firstLeg * secondLeg / 2;
         return area;
     }
